ThrowHelper use changes
